pfs-49: Moved linting to the golangci-lint-action
