+++ finance-hub/internal/api/api_client_test.go
@@ -3,13 +3,27 @@ package api
 import (
 	"context"
 	"github.com/ministryofjustice/opg-go-common/logging"
-	"github.com/opg-sirius-finance-hub/finance-hub/internal/mocks"
 	"net/http"
 	"os"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
 )
+
+// MockClient is the mock client
+type MockClient struct {
+	DoFunc func(req *http.Request) (*http.Response, error)
+}
+
+var (
+	// GetDoFunc fetches the mock client's `Do` func
+	GetDoFunc func(req *http.Request) (*http.Response, error)
+)
+
+// Do is the mock client's `Do` func
+func (m *MockClient) Do(req *http.Request) (*http.Response, error) {
+	return GetDoFunc(req)
+}
 
 func getContext(cookies []*http.Cookie) Context {
 	return Context{
@@ -42,8 +56,8 @@ func TestStatusError(t *testing.T) {
 	assert.Equal(t, err, err.Data())
 }
 
-func SetUpTest() (*logging.Logger, *mocks.MockClient) {
+func SetUpTest() (*logging.Logger, *MockClient) {
 	logger := logging.New(os.Stdout, "opg-sirius-finance-hub")
-	mockClient := &mocks.MockClient{}
+	mockClient := &MockClient{}
 	return logger, mockClient
 }
