+++ finance-hub/internal/api/get_account_information_test.go
@@ -2,7 +2,6 @@ package api
 
 import (
 	"bytes"
-	"github.com/opg-sirius-finance-hub/finance-hub/internal/mocks"
 	"github.com/opg-sirius-finance-hub/shared"
 	"io"
 	"net/http"
@@ -12,20 +11,19 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-func TestGetHeaderAccountDetails(t *testing.T) {
+func TestGetAccountInformation(t *testing.T) {
 	logger, mockClient := SetUpTest()
 	client, _ := NewApiClient(mockClient, "http://localhost:3000", "http://localhost:8181", logger)
 
 	json := `{
-            "clientId": 2,
-            "cachedOutstandingBalance": "22.22",
-            "cachedCreditBalance": "1.01",
-            "paymentMethod": "Demanded"
+            "outstandingBalance": 2222,
+            "creditBalance": 101,
+            "paymentMethod": "DEMANDED"
         }`
 
 	r := io.NopCloser(bytes.NewReader([]byte(json)))
 
-	mocks.GetDoFunc = func(*http.Request) (*http.Response, error) {
+	GetDoFunc = func(*http.Request) (*http.Response, error) {
 		return &http.Response{
 			StatusCode: 200,
 			Body:       r,
@@ -35,7 +33,7 @@ func TestGetHeaderAccountDetails(t *testing.T) {
 	expectedResponse := shared.AccountInformation{
 		OutstandingBalance: 2222,
 		CreditBalance:      101,
-		PaymentMethod:      "Demanded",
+		PaymentMethod:      "DEMANDED",
 	}
 
 	headerDetails, err := client.GetAccountInformation(getContext(nil), 2)
@@ -43,7 +41,7 @@ func TestGetHeaderAccountDetails(t *testing.T) {
 	assert.Equal(t, nil, err)
 }
 
-func TestGetHeaderAccountDetailsReturnsUnauthorisedClientError(t *testing.T) {
+func TestGetAccountInformationReturnsUnauthorisedClientError(t *testing.T) {
 	logger, _ := SetUpTest()
 	svr := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		w.WriteHeader(http.StatusUnauthorized)
@@ -55,7 +53,7 @@ func TestGetHeaderAccountDetailsReturnsUnauthorisedClientError(t *testing.T) {
 	assert.Equal(t, ErrUnauthorized, err)
 }
 
-func TestHeaderAccountDetailsReturns500Error(t *testing.T) {
+func TestAccountInformationReturns500Error(t *testing.T) {
 	logger, _ := SetUpTest()
 	svr := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		w.WriteHeader(http.StatusInternalServerError)
