--- finance-hub/internal/mocks/client.go
@@ -1,18 +1,0 @@
-package mocks
-
-import "net/http"
-
-// MockClient is the mock client
-type MockClient struct {
-	DoFunc func(req *http.Request) (*http.Response, error)
-}
-
-var (
-	// GetDoFunc fetches the mock client's `Do` func
-	GetDoFunc func(req *http.Request) (*http.Response, error)
-)
-
-// Do is the mock client's `Do` func
-func (m *MockClient) Do(req *http.Request) (*http.Response, error) {
-	return GetDoFunc(req)
-}
